Dried update mutations in GraphQL
